+++ website/src/app/shell-wordings.d.ts
@@ -3,7 +3,9 @@ import { Language } from "./languages";
 export interface ShellWording {
   navLinks: Array<string>;
   languageSwitch: string;
+  languageSwitcherLabel: string;
   editButton: string;
+  copyright: string;
 }
 
 export type ShellWordings = Record<Language, ShellWording>;